tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation_in_document_style.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation specified in document stylesheet</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animations">
   <meta name="flags" content="dom">
   <meta name="assert" content="An animation specified in the document
   stylesheet will begin at document load">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         from {
            margin-left: 100px;
         }
         to {
            margin-left: 0;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         animation: slide 1s;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var divMargin = getComputedStyle(div, null).marginLeft;
         var passed = parseFloat(divMargin) > 90;
         test(function(){assert_true(passed);}, testName); done();
         log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should immediately start moving left, stopping after
   one second.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
